All Error Fixed 336 Warning Left To Resolve

+++ ACL/src/ACL.Infrastructure/Repositories/V1/AclCompanyModuleRepository.cs
@@ -133,7 +133,7 @@ namespace ACL.Infrastructure.Repositories.V1
             }
         }
         /// <inheritdoc/>
-        public AclCompanyModule PrepareInputData(AclCompanyModuleRequest request, ulong Id = 0, AclCompanyModule _aclCompanyModule = null)
+        public AclCompanyModule PrepareInputData(AclCompanyModuleRequest request, ulong Id = 0, AclCompanyModule? _aclCompanyModule = null)
         {
             bool valid = IsValidForCreateOrUpdate(request.CompanyId, request.ModuleId);
             AclCompanyModule aclCompanyModule = new AclCompanyModule();
@@ -235,6 +235,7 @@ namespace ACL.Infrastructure.Repositories.V1
             try
             {
                 var delete = _dbContext.AclCompanyModules.Find(id);
+                if (delete != null)
                 _dbContext.AclCompanyModules.Remove(delete);
                 _dbContext.SaveChangesAsync();
                 return delete;
